+++ src/index.ts
@@ -110,3 +110,6 @@ export {
     StarterKitPreset,
     MarkedPreset
 } from "./presets";
+
+// React wrapper
+export * from './react';